+++ ltv_mpc/exceptions.py
@@ -20,5 +20,8 @@
 class LTVMPCException(Exception):
     """Base class for exceptions from ltv-mpc."""
 
-class ProblemDefinitionError(Exception):
+class ProblemDefinitionError(LTVMPCException):
     """Problem definition is incorrect."""
+
+class StateError(LTVMPCException):
+    """State is incorrect."""